requirement.txt
fastapi==0.111.0
uvicorn==0.15.0
boto3==1.34.137
passlib==1.7.4
sqlalchemy==2.0.31
sqlmodel==0.0.19
psycopg2-binary==2.9.9
asyncpg==0.27.0